Refactor steps
This organises the Cucumber steps into more logical files to make them
easier to keep track of.
