+++ lib/engines/content_block_manager/features/step_definitions/schema_steps.rb
@@ -1,0 +1,31 @@
+Given("a schema {string} exists with the following fields:") do |block_type, table|
+  fields = table.hashes
+  @schemas ||= {}
+  body = {
+    "type" => "object",
+    "required" => fields.select { |f| f["required"] == "true" }.map { |f| f["field"] },
+    "additionalProperties" => false,
+    "properties" => fields.map { |f|
+      [f["field"], { "type" => f["type"], "format" => f["format"] }]
+    }.to_h,
+  }
+  @schemas[block_type] = build(:content_block_schema, block_type:, body:)
+  ContentBlockManager::ContentBlock::Schema.stubs(:all).returns(@schemas.values)
+end
+
+When("I click on the {string} schema") do |schema_id|
+  @schema = @schemas[schema_id]
+  ContentBlockManager::ContentBlock::Schema.expects(:find_by_block_type).with(schema_id).at_least_once.returns(@schema)
+  choose @schema.name
+  click_save_and_continue
+end
+
+Then("I should see a form for the schema") do
+  expect(page).to have_content(@schema.name)
+end
+
+Then("I should see all the schemas listed") do
+  @schemas.values.each do |schema|
+    expect(page).to have_content(schema.name)
+  end
+end
