+++ lib/engines/content_block_manager/features/step_definitions/timeline_steps.rb
@@ -1,0 +1,18 @@
+Then("I should see the created event on the timeline") do
+  expect(page).to have_selector(".timeline__title", text: "Email address created")
+  expect(page).to have_selector(".timeline__byline", text: "by #{@user.name}")
+end
+
+Then(/^I should see ([^"]*) publish events on the timeline$/) do |count|
+  expect(page).to have_selector(".timeline__title", text: "Email address published", count:)
+end
+
+Then("I should see the publish event on the timeline") do
+  expect(page).to have_selector(".timeline__title", text: "Email address published")
+  expect(page).to have_selector(".timeline__byline", text: "by Scheduled Publishing Robot")
+end
+
+Then("I should see the scheduled event on the timeline") do
+  expect(page).to have_selector(".timeline__title", text: "Email address scheduled")
+  expect(page).to have_selector(".timeline__byline", text: "by #{@user.name}")
+end
